Uploaded the user stories and case study
I also cleaned up the user stories to make the wording more consistent (e.g. replaced all instances of "non-member", "potential-member" etc. with "user"), as well as adding asterisks around all stories which aren't directly mentioned in the case study requirements. My reasoning for this is so that we can prioritise the ones which we absolutely need, and then work on everything which would be good to have but isn't technically required.